[ingest] fix DIGREPO-629
Previously the metadata and data for an ETD were being matched up by
trying to make sure index `n` of metadata and data arrays corresponded
to metadata and data for the same object.  That was hard to guarantee
even when fetching metadata at the time of ingest, and it wasn't
generally true at all with prefetched metadata.

So instead compare PDF filenames to MARC 956.f fields to match up the
metadata and data.