Add full text searching for PDFs
This commit introduces two classes that can be used
to index full text content of PDFs intro Solr.

There is a `FullTextExtractor` class and a
`FullTextToSolrJob` class. The first one handles
getting the the extracted text from Solr by posting
the PDF data to Solr. The job class gets the full text
and updates the Solr index with the text.

To run the job:

```ruby
FullTextToSolrJob.new("fk4jq25q2f", Image.first.file_sets[0].files.first.content).perform
```

The first param is the `SolrDocument` id, and the second param is a PDF datastream
in an ActiveFedora object.