update parser version
From Jenkins
```
Your bundle is locked to parser (2.5.0.4), but that version could not be found
in any of the sources listed in your Gemfile. If you haven't changed sources,
that means the author of parser (2.5.0.4) has removed it. You'll need to update
your bundle to a different version of parser (2.5.0.4) that hasn't been removed
in order to install.
```